requirements.txt
streamlit
psycopg2 
python-dotenv
pandas
fpdf
opencv-python
pyzbar
smtplib
